requirements_mkdocs.txt
Babel==2.15.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ghp-import==2.1.0
idna==3.7
install==1.3.5
Jinja2==3.1.4
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-drawio==1.6.2
mkdocs-get-deps==0.2.0
mkdocs-macros-plugin==1.0.5
mkdocs-material==9.5.28
mkdocs-material-extensions==1.3.1
mkdocs-nav-weight==0.2.0
mkdocs-redirects==1.2.1
mkdocs-table-reader-plugin==2.2.2
numpy==2.0.0
packaging==24.1
paginate==0.5.6
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
Pygments==2.18.0
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
termcolor==2.4.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
